Correções finais antes de mais alguns testes

+++ API/BlogTJMT.Data/Repositories/PostRepository.cs
@@ -1,4 +1,6 @@
-﻿using BlogTJMT.Common.Validations;
+﻿using BlogTJMT.Common.Enum;
+using BlogTJMT.Common.Resources;
+using BlogTJMT.Common.Validations;
 using BlogTJMT.Data.DataContexts;
 using BlogTJMT.Domain.Contract.Repositories;
 using BlogTJMT.Domain.Model;
@@ -23,44 +25,56 @@ namespace BlogTJMT.Data.Repositories
             _db.SaveChanges();
 
             RemoveComentarios(id);
-            RemoveVinculosCategoria(id);
-        }
-
-        private void RemoveVinculosCategoria(int id)
-        {
-            var result = _db.PostCategorias.FirstOrDefault(coluna => coluna.Post.Id == id);
-            _db.PostCategorias.Remove(result);
-            _db.SaveChanges();
         }
 
         private void RemoveComentarios(int id)
         {
             var result = _db.PostComentarios.FirstOrDefault(coluna => coluna.Post.Id == id);
-            _db.PostComentarios.Remove(result);
-            _db.SaveChanges();
+            if (result != null)
+            {
+                _db.PostComentarios.Remove(result);
+                _db.SaveChanges();
+            }
         }
 
         public void Dispose() => _db.Dispose();
         public List<Post> Get()
         {
-            return _db.Posts.Include(nameof(Usuario)).Include($"{nameof(Usuario)}.{nameof(Perfil)}").ToList();
+            return _db.Posts
+                        .Include(nameof(Usuario))
+                        .Include(nameof(Categoria))
+                        .ToList();
         }
 
         public List<Post> Get(string titulo)
         {
             return _db.Posts
                         .Include(nameof(Usuario))
-                        .Include($"{nameof(Usuario)}.{nameof(Perfil)}")
+                        .Include(nameof(Categoria))
                         .Where(coluna => coluna.Titulo.Contains(titulo)).ToList();
         }
 
-        public Post Get(int id) { return _db.Posts.Include(nameof(Usuario)).Include($"{nameof(Usuario)}.{nameof(Perfil)}").FirstOrDefault(coluna => coluna.Id == id); }
+        public Post Get(int id)
+        {
+            return _db.Posts
+                        .Include(nameof(Usuario))
+                        .Include(nameof(Categoria))
+                        .FirstOrDefault(coluna => coluna.Id == id);
+        }
+
+        public List<Post> GetTop5()
+        {
+            return _db.Posts
+                        .Include(nameof(Usuario))
+                        .Include(nameof(Categoria))
+                        .OrderByDescending(coluna => coluna.Curtidas).ThenBy(coluna => coluna.Data).Take(5).ToList();
+        }
 
         public List<Post> GetMaisCurtidas()
         {
             var result = _db.Posts
                                 .Include(nameof(Usuario))
-                                .Include($"{nameof(Usuario)}.{nameof(Perfil)}")
+                                .Include(nameof(Categoria))
                                 .ToList().OrderBy(coluna => coluna.Curtidas);
             return result.ToList();
         }
@@ -69,7 +83,7 @@ namespace BlogTJMT.Data.Repositories
         {
             var result = _db.Posts
                                 .Include(nameof(Usuario))
-                                .Include($"{nameof(Usuario)}.{nameof(Perfil)}")
+                                .Include(nameof(Categoria))
                                 .ToList().OrderBy(coluna => coluna.Visualizacoes);
             return result.ToList();
         }
@@ -77,28 +91,47 @@ namespace BlogTJMT.Data.Repositories
         public Post Post(Post post)
         {
             ValidationClass.ValidaClasse(post);
+
+            ValidaUsuario(post.Usuario);
+            post.Usuario = null;
+
             _db.Posts.Add(post);
             _db.SaveChanges();
-
-            var postCategoria = new PostCategoria();
-            post.Categorias.ForEach(campo => SalvaPostCategoria(campo, post));
 
             return post;
         }
 
-        private void SalvaPostCategoria(Categoria categoria, Post post)
+        private void ValidaUsuario(Usuario usuario)
         {
-            _db.PostCategorias.Add(new PostCategoria
-            {
-                CategoriaId = categoria.Id,
-                PostId = post.Id
-            });
-            _db.SaveChanges();
+            ValidationClass.ValidaClasse(usuario);
+
+            var result = (from item in _db.Usuarios
+                          where item.Email.ToLower() == (usuario.Email.ToLower()) && item.Senha == (usuario.Senha)
+                          select item).FirstOrDefault();
+
+            if (result == null)
+                throw new Exception(MensagensErro.PostSemUsuario);
+
+            ValidaPermicaoUsuario(result);
+        }
+
+        private void ValidaPermicaoUsuario(Usuario usuario)
+        {
+            var result = (from item in _db.PerfilPermicoes.Include(nameof(Permicao))
+                          where item.PerfilId == usuario.PerfilId
+                          select item).FirstOrDefault();
+
+            if (!(result.Permicao.Enum == PermicaoEnum.Admin || result.Permicao.Enum == PermicaoEnum.Editor))
+                throw new Exception(MensagensErro.UsuarioSemPermicao);
         }
 
         public Post Put(Post post)
         {
             ValidationClass.ValidaClasse(post);
+
+            ValidaUsuario(post.Usuario);
+            post.Usuario = null;
+
             _db.Entry(post).State = System.Data.Entity.EntityState.Modified;
             _db.SaveChanges();
